fix numpy float compatibility in test_quick_focus_solve_apply for pos2 and pos1

+++ tests/test_solves.py
@@ -85,8 +85,8 @@ class TestQuickfocusSolve:
         solve.apply()
 
         # Check that surface has been shifted
-        pos2 = optic.surface_group.positions[-1][0]
-        pos1 = optic.surface_group.positions[-2][0]
+        pos2 = optic.surface_group.positions[-1].astype(float)[0]
+        pos1 = optic.surface_group.positions[-2].astype(float)[0]
 
         assert pos2 - pos1 == pytest.approx(thickness)
